Move the kafkasource client.go into separate package
To prevent informer injection happening whenever another package uses this code
