+++ pkg/source/client/client.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package source
+package client
 
 import (
 	"context"
